Move _test_print_widgets to utils
